requirements.txt
Flask
requests
requests_oauthlib
psycopg2-binary
python-telegram-bot==13.15
pyTelegramBotAPI==4.14.0
